updated external_gui components for new structure
