+++ pylabnet/gui/pyqt/external_gui.py
@@ -509,175 +509,6 @@ class Window(QtWidgets.QMainWindow):
         )
 
 
-class Service(ServiceBase):
-
-    def exposed_assign_plot(self, plot_widget, plot_label, legend_widget):
-        return self._module.assign_plot(
-            plot_widget=plot_widget,
-            plot_label=plot_label,
-            legend_widget=legend_widget
-        )
-
-    def exposed_clear_plot(self, plot_widget):
-        return self._module.clear_plot(
-            plot_widget=plot_widget
-        )
-
-    def exposed_assign_curve(self, plot_label, curve_label, error=False):
-        return self._module.assign_curve(
-            plot_label=plot_label,
-            curve_label=curve_label,
-            error=error
-        )
-
-    def exposed_remove_curve(self, plot_label, curve_label):
-        self._module.remove_curve(
-            plot_label=plot_label,
-            curve_label=curve_label
-        )
-
-    def exposed_assign_scalar(self, scalar_widget, scalar_label):
-        return self._module.assign_scalar(
-            scalar_widget=scalar_widget,
-            scalar_label=scalar_label
-        )
-
-    def exposed_assign_label(self, label_widget, label_label):
-        return self._module.assign_label(
-            label_widget=label_widget,
-            label_label=label_label
-        )
-
-    def exposed_assign_event_button(self, event_widget, event_label):
-        return self._module.assign_event_button(
-            event_widget=event_widget,
-            event_label=event_label,
-
-        )
-
-    def exposed_set_curve_data(self, data_pickle, plot_label, curve_label, error_pickle=None):
-        data = pickle.loads(data_pickle)
-        error = pickle.loads(error_pickle)
-        return self._module.set_curve_data(
-            data=data,
-            plot_label=plot_label,
-            curve_label=curve_label,
-            error=error
-        )
-
-    def exposed_set_scalar(self, value_pickle, scalar_label):
-        value = pickle.loads(value_pickle)
-        return self._module.set_scalar(
-            value=value,
-            scalar_label=scalar_label
-        )
-
-    def exposed_get_scalar(self, scalar_label):
-        return pickle.dumps(self._module.get_scalar(scalar_label))
-
-    def exposed_activate_scalar(self, scalar_label):
-        return self._module.activate_scalar(scalar_label)
-
-    def exposed_deactivate_scalar(self, scalar_label):
-        return self._module.deactivate_scalar(scalar_label)
-
-    def exposed_set_label(self, text, label_label):
-        return self._module.set_label(
-            text=text,
-            label_label=label_label
-        )
-
-    def exposed_was_button_pressed(self, event_label):
-        return self._module.was_button_pressed(event_label)
-
-    def exposed_change_button_background_color(self, event_label, color):
-        return self._module.change_button_background_color(self, event_label, color)
-
-
-class Client(ClientBase):
-
-    def assign_plot(self, plot_widget, plot_label, legend_widget):
-        return self._service.exposed_assign_plot(
-            plot_widget=plot_widget,
-            plot_label=plot_label,
-            legend_widget=legend_widget
-        )
-
-    def clear_plot(self, plot_widget):
-        return self._service.exposed_clear_plot(
-            plot_widget=plot_widget
-        )
-
-    def assign_curve(self, plot_label, curve_label, error=False):
-        return self._service.exposed_assign_curve(
-            plot_label=plot_label,
-            curve_label=curve_label,
-            error=error
-        )
-
-    def remove_curve(self, plot_label, curve_label):
-        return self._service.exposed_remove_curve(
-            plot_label=plot_label,
-            curve_label=curve_label
-        )
-
-    def assign_scalar(self, scalar_widget, scalar_label):
-        self._service.exposed_assign_scalar(
-            scalar_widget=scalar_widget,
-            scalar_label=scalar_label
-        )
-
-    def assign_label(self, label_widget, label_label):
-        return self._service.exposed_assign_label(
-            label_widget=label_widget,
-            label_label=label_label
-        )
-
-    def assign_event_button(self, event_widget, event_label):
-        return self._service.exposed_assign_event_button(
-            event_widget=event_widget,
-            event_label=event_label,
-        )
-
-    def set_curve_data(self, data, plot_label, curve_label, error=None):
-        data_pickle = pickle.dumps(data)
-        error_pickle = pickle.dumps(error)
-        return self._service.exposed_set_curve_data(
-            data_pickle=data_pickle,
-            plot_label=plot_label,
-            curve_label=curve_label,
-            error_pickle=error_pickle
-        )
-
-    def set_scalar(self, value, scalar_label):
-        value_pickle = pickle.dumps(value)
-        return self._service.exposed_set_scalar(
-            value_pickle=value_pickle,
-            scalar_label=scalar_label
-        )
-
-    def get_scalar(self, scalar_label):
-        return pickle.loads(self._service.exposed_get_scalar(scalar_label))
-
-    def activate_scalar(self, scalar_label):
-        return self._service.exposed_activate_scalar(scalar_label)
-
-    def deactivate_scalar(self, scalar_label):
-        return self._service.exposed_deactivate_scalar(scalar_label)
-
-    def set_label(self, text, label_label):
-        return self._service.exposed_set_label(
-            text=text,
-            label_label=label_label
-        )
-
-    def was_button_pressed(self, event_label):
-        return self._service.exposed_was_button_pressed(event_label)
-
-    def change_button_background_color(self, event_label, color):
-        return self._service.exposed_change_button_background_color(self, event_label, color)
-
-
 class Plot:
     """ Class for plot widgets inside of a Window
 
@@ -1002,48 +833,3 @@ class EventButton:
         result = copy.deepcopy(self.was_pushed)
         self.reset_button()
         return result
-
-
-def launch(logger=None, port=None, name=None):
-    """ Instantiates a default main window + server
-
-    :param logger: (LogClient) instance of LogClient for logging purposes
-    :param port: (int) port number for the GUI server
-    :param name: (str) name of server to display
-    """
-
-    # Create app and instantiate main window
-    app = QtWidgets.QApplication(sys.argv)
-    ui = input('Please enter the .ui file to use as a template:\n>> ')
-    try:
-        main_window = Window(app, gui_template=ui)
-    except FileNotFoundError:
-        print('Could not find .ui file, '
-              'please check that it is in the pylabnet/gui/pyqt/gui_templates directory')
-        raise
-    gui_service = Service()
-    gui_service.assign_module(module=main_window)
-    gui_service.assign_logger(logger=logger)
-
-    if port is None:
-        port = np.random.randint(1, 9999)
-    gui_server = GenericServer(
-        service=gui_service,
-        host='localhost',
-        port=port
-    )
-
-    gui_server.start()
-
-    # Update GUI with server-specific details
-    main_window.ip_label.setText('IP Address: {}'.format(
-        socket.gethostbyname(socket.gethostname())
-    ))
-    main_window.port_label.setText('Port: {}'.format(port))
-
-    # Run the GUI until the stop button is clicked
-    while not main_window.stop_button.isChecked():
-        main_window.configure_widgets()
-        main_window.update_widgets()
-        main_window.force_update()
-    sys.exit(app.exec_())
